update MyTeamList Information
