+++ src/views/myTeam/MyTeamPage.vue
@@ -6,33 +6,39 @@
             <v-tab>팀기록 조회</v-tab>
         </v-tabs>
         <router-view :teamSeq="teamId"/>
+        <!-- <h2>{{$route.params.teamSeq}}</h2>
+        <h2>{{teamSeq}}</h2> -->
     </div>
 </template>
 
 <script>
     //import { API } from '@/api/MyTeamAPI.js';
     export default {
-        mounted (){
-            this.onClick();
-        },
+        data: () => {
+            return {
+                // teamId: 4,
+                teamId: undefined
+            }
+        },        
         props: {
             teamSeq: {
                 type: Number,
                 required: true
             }
         },
-        data: () => {
-            return {
-                // teamId: this.teamSeq
-                teamId: 4
-            }
-        },
         methods: {
             onClick () {
                 // console.log(123);
-                console.log(this.teamSeq);
+                // console.log(this.teamSeq);
+            },
+            init(){
+                this.teamId = this.teamSeq
             }
-        }
+        },
+        mounted (){
+            // this.onClick();
+            this.init();
+        }        
     }
 </script>
